get top k peaks
