+++ backend/app.py
@@ -1,18 +1,13 @@
 import os
 import requests
 import urllib.parse
-import json
-
 from Article import Article
 from newspaper import Article as Article3k
-
 from flask import Flask, jsonify, request
-# from dotenv import load_dotenv
+from dotenv import load_dotenv
 from flask_cors import CORS
 
-from pytrends.request import TrendReq
-
-# load_dotenv()
+load_dotenv()
 
 app = Flask(__name__)
 cors = CORS(app)
@@ -47,7 +42,7 @@ def get_news_articles():
   encoded_query = urllib.parse.urlencode(query)
   response = requests.get(request_url, params=encoded_query)
   parsed_articles = parse_articles(response.json())
-  #TODO: error handlings
+  # TODO: error handlings
   #now, use this list of articles to make call to scrape the article
   #return "good request!"
   return response.json(), 200
@@ -76,24 +71,5 @@ def calculate_sentiment(url):
     return 0.2
 
 
-def get_google_trends(keywords, dates):
-    """
-
-    :param keywords:
-    :param dates: 'YYYY-MM-DD, YYYY_MM-DD'
-    :return:
-    """
-
-    timeframe = dates.replace(',', ' ')
-    print(timeframe)
-
-    py_trends = TrendReq(hl='en-US', tz=360)
-    py_trends.build_payload(keywords, cat=0, timeframe=timeframe, geo='', gprop='')
-    trends_df = py_trends.interest_over_time()
-    trends_df = trends_df.drop(columns='isPartial')
-    return trends_df.to_json()
-
-
 if __name__ == '__main__':
-    print(get_google_trends(['bitcoin'], '2022-01-01,2022-01-12'))
-    # app.run(host='0.0.0.0', port=5000)
+    app.run(host='0.0.0.0', port=5000)
