+++ backend/utils.py
@@ -1,0 +1,2 @@
+def timestamp_to_string(timestamp):
+    return timestamp.strftime('%Y-%m-%d')
